+++ src/index.hpp
@@ -136,7 +136,7 @@ struct StrobemerIndex {
         if (is_filtered_forward(position_revcomp)) {
             return true;
         }
-        size_t count = get_count_full(position) + get_count_full(position_revcomp);
+        size_t count = get_count_full_forward(position) + get_count_full_forward(position_revcomp);
 
         return count > filter_cutoff;
     }
@@ -148,19 +148,7 @@ struct StrobemerIndex {
 
     bool is_partial_filtered(bucket_index_t position, randstrobe_hash_t hash_revcomp) const {
         assert(position < randstrobes.size());
-        if (is_partial_filtered_forward(position)) {
-            return true;
-        }
-        bucket_index_t position_revcomp = find_partial(hash_revcomp);
-        if (position_revcomp == end()) {
-            return false;
-        }
-        if (is_partial_filtered_forward(position_revcomp)) {
-            return true;
-        }
-        size_t count = get_count_partial(position) + get_count_partial(position_revcomp);
-
-        return count > filter_cutoff;
+        return is_partial_filtered_forward(position);
     }
 
     unsigned int get_strobe1_position(bucket_index_t position) const {
@@ -189,7 +177,17 @@ struct StrobemerIndex {
         return randstrobes.size();
     }
 
-    unsigned int get_count_full(bucket_index_t position) const {
+    /* Count number of hits for the randstrobe *and* its "reverse complement" */
+    unsigned int get_count_full(bucket_index_t position, randstrobe_hash_t hash_revcomp) const {
+        size_t reverse_count{0};
+        bucket_index_t position_revcomp = find_full(hash_revcomp);
+        if (position_revcomp != end()) {
+            reverse_count = get_count_full_forward(position_revcomp);
+        }
+        return reverse_count + get_count_full_forward(position);
+    }
+
+    unsigned int get_count_full_forward(bucket_index_t position) const {
         return get_count(position, RANDSTROBE_HASH_MASK);
     }